binaryen turned off for win32

+++ src/cli/post-install.ts
@@ -18,7 +18,7 @@ const PLATFORM = os.platform();
 const ARCH = os.arch();
 console.log(`Current platform: ${PLATFORM}, current architecture: ${ARCH}`);
 
-const SUPPORTED_PLATFORMS = ["linux", "darwin"]; // Unsaported platforms: 'win32', 'aix', 'freebsd', 'openbsd', 'sunos', 'android'
+const SUPPORTED_PLATFORMS = ["linux", "darwin", "win32"]; // Unsaported platforms: 'aix', 'freebsd', 'openbsd', 'sunos', 'android'
 const SUPPORTED_ARCH = ["x64", "arm64"]; // Unsaported arch: 'arm', 'ia32', 'mips','mipsel', 'ppc', 'ppc64', 's390', 's390x', 'x32'
 
 if (!SUPPORTED_PLATFORMS.includes(PLATFORM)) {
@@ -31,30 +31,31 @@ if (!SUPPORTED_ARCH.includes(ARCH)) {
   process.exit(1);
 }
 
-signale.await("Installing wasi-stub...");
+if (PLATFORM !== "win32") {
+  signale.await("Installing wasi-stub...");
 
-const BINARYEN_VERSION = `0.1.10`;
-const BINARYEN_VERSION_TAG = `v${BINARYEN_VERSION}`;
+  const BINARYEN_VERSION = `0.1.10`;
+  const BINARYEN_VERSION_TAG = `v${BINARYEN_VERSION}`;
 
-const BINARYEN_SYSTEM_NAME =
-  PLATFORM === "linux" ? "Linux" :
-  PLATFORM === "darwin" ? "macOS" :
-  PLATFORM === "win32" ? "windows" : "other";
+  const BINARYEN_SYSTEM_NAME =
+    PLATFORM === "linux" ? "Linux" :
+      PLATFORM === "darwin" ? "macOS" : "other";
 
-const BINARYEN_ARCH_NAME =
-  ARCH === "x64" ? "X64" :
-  ARCH === "arm64" ? "arm64" : "other";
+  const BINARYEN_ARCH_NAME =
+    ARCH === "x64" ? "X64" :
+      ARCH === "arm64" ? "arm64" : "other";
 
-const BINARYEN_TAR_NAME = `binaryen-${BINARYEN_SYSTEM_NAME}-${BINARYEN_ARCH_NAME}.tar.gz`;
+  const BINARYEN_TAR_NAME = `binaryen-${BINARYEN_SYSTEM_NAME}-${BINARYEN_ARCH_NAME}.tar.gz`;
 
-await download(
-  `https://github.com/near/binaryen/releases/download/${BINARYEN_VERSION_TAG}/${BINARYEN_TAR_NAME}`
-);
+  await download(
+    `https://github.com/near/binaryen/releases/download/${BINARYEN_VERSION_TAG}/${BINARYEN_TAR_NAME}`
+  );
 
-fs.mkdirSync("binaryen");
+  fs.mkdirSync("binaryen");
 
-await executeCommand(`tar xvf ${BINARYEN_TAR_NAME} --directory binaryen`);
-fs.rmSync(BINARYEN_TAR_NAME);
+  await executeCommand(`tar xvf ${BINARYEN_TAR_NAME} --directory binaryen`);
+  fs.rmSync(BINARYEN_TAR_NAME);
+}
 
 signale.await("Installing QuickJS...");
 
@@ -63,12 +64,12 @@ const QUICK_JS_VERSION_TAG = `v${QUICK_JS_VERSION}`;
 
 const QUICK_JS_SYSTEM_NAME =
   PLATFORM === "linux" ? "Linux" :
-  PLATFORM === "darwin"? "macOS":
-  PLATFORM === "win32" ? "windows" : "other";
+    PLATFORM === "darwin" ? "macOS" :
+      PLATFORM === "win32" ? "windows" : "other";
 
 const QUICK_JS_ARCH_NAME =
   ARCH === "x64" ? "X64" :
-  ARCH === "arm64" ? "arm64" : "other";
+    ARCH === "arm64" ? "arm64" : "other";
 
 const QUICK_JS_TAR_NAME = `${QUICK_JS_VERSION_TAG}.tar.gz`;
 const QUICK_JS_DOWNLOADED_FOLDER_NAME = `quickjs-${QUICK_JS_VERSION}`;
@@ -106,8 +107,8 @@ const WASI_SDK_MINOR_VER = 0;
 const WASI_SDK_DOWNLOADED_FOLDER_NAME = `wasi-sdk-${WASI_SDK_MAJOR_VER}.${WASI_SDK_MINOR_VER}`;
 const WASI_SDK_SYSTEM_NAME =
   PLATFORM === "linux" ? "linux" :
-  PLATFORM === "darwin" ? "macos" :
-  PLATFORM === "win32" ? "windows" : "other";
+    PLATFORM === "darwin" ? "macos" :
+      PLATFORM === "win32" ? "windows" : "other";
 const WASI_SDK_TAR_NAME = `${WASI_SDK_DOWNLOADED_FOLDER_NAME}-${WASI_SDK_SYSTEM_NAME}.tar.gz`;
 
 // Download WASI SDK
